run and wait
