+++ frontend/docs/pages/home/_meta.js
@@ -24,10 +24,17 @@ export default {
   "scheduled-runs": "Scheduled Trigger",
   "cron-runs": "Cron Trigger",
   "run-on-event": "Event Trigger",
+  "bulk-run": "Bulk Run Many",
   "streaming-results": {
     "title": "Streaming Results",
     "display": "hidden"
   },
+  "--flow-control": {
+    "title": "Flow Control",
+    "type": "separator"
+  },
+  "concurrency": "Concurrency",
+  "rate-limits": "Rate Limits",
   "running-from-external-services": {
     "title": "Running from External Services",
     "display": "hidden"
@@ -78,12 +85,6 @@ export default {
   },
   "docker": "Running with Docker",
   "worker-healthchecks": "Worker Health Checks",
-  "--flow-control": {
-    "title": "Flow Control",
-    "type": "separator"
-  },
-  "concurrency": "Concurrency",
-  "rate-limits": "Rate Limits",
   "--cancellation": {
     "title": "Cancellation",
     "type": "separator",
