+++ frontend/docs/pages/home/bulk-run.mdx
@@ -44,18 +44,34 @@ You can also `await` the result of `aio_run`:
 
     <GithubSnippet src={SimplePy} target="Running a Task AIO" />
 
+    Since the `aio_run` method returns a coroutine, you can spawn multiple tasks in parallel and await using `asyncio.gather`.
+    <GithubSnippet src={SimplePy} target="Running Multiple Tasks" />
+
+
 Note that the type of `input` here is a Pydantic model that matches the input schema of your workflow.
 
   </Tabs.Tab>
   <Tabs.Tab title="Typescript">
   You can use your `Task` object to run a task and wait for it to complete by calling the `run` method. This method will return a promise that resolves when the task completes and returns the result.
     <GithubSnippet src={SimpleTs} target="Running a Task" />
+
+    Since the `run` method returns a promise, you can spawn multiple tasks in parallel and await using `Promise.all`.
+    <GithubSnippet src={SimpleTs} target="Running Multiple Tasks" />
   </Tabs.Tab>
   <Tabs.Tab title="Go">
     You can use your `Task` object to run a task and wait for it to complete by calling the `Run` method. This method will block until the task completes and return the result.
     <GithubSnippet src={SimpleGo} target="Running a Task" />
+
+    You can also run multiple tasks in parallel by calling `Run` multiple times in goroutines and using a `sync.WaitGroup` to wait for them to complete.
+    <GithubSnippet src={SimpleGo} target="Running Multiple Tasks" />
+    
   </Tabs.Tab>
 </UniversalTabs>
+
+
+<Callout type="info">
+  While you can run multiple tasks in parallel using the `Run` method, this is not recommended for large numbers of tasks. Instead, we recommend using [batch methods](./batching.mdx) large parallel task execution.
+</Callout>
 
 ## Spawning Tasks from within a Task
 
@@ -84,35 +100,3 @@ And that's it! The parent task will run and spawn the child task, and then will 
     <GithubSnippet src={ParentTaskGo} target="Spawning Tasks from within a Task" />
   </Tabs.Tab>
 </UniversalTabs>
-
-## Running Tasks in Parallel
-
-Sometimes you may want to run multiple tasks concurrently. Here's how to do that in each language:
-
-<UniversalTabs items={["Python", "Typescript", "Go"]}>
-<Tabs.Tab title="Python">
-
-Since the `aio_run` method returns a coroutine, you can spawn multiple tasks in parallel and await using `asyncio.gather`.
-
-<GithubSnippet src={SimplePy} target="Running Multiple Tasks" />
-
-</Tabs.Tab>
-<Tabs.Tab title="Typescript">
-
-Since the `run` method returns a promise, you can spawn multiple tasks in parallel and await using `Promise.all`.
-
-<GithubSnippet src={SimpleTs} target="Running Multiple Tasks" />
-
-</Tabs.Tab>
-<Tabs.Tab title="Go">
-
-You can run multiple tasks in parallel by calling `Run` multiple times in goroutines and using a `sync.WaitGroup` to wait for them to complete.
-
-<GithubSnippet src={SimpleGo} target="Running Multiple Tasks" />
-
-</Tabs.Tab>
-</UniversalTabs>
-
-<Callout type="info">
-  While you can run multiple tasks in parallel using the `Run` method, this is not recommended for large numbers of tasks. Instead, we recommend using [bulk run methods](./bulk-run.mdx) for large parallel task execution.
-</Callout>
