+++ sdks/python/examples/fanout/worker.py
@@ -9,8 +9,6 @@ hatchet = Hatchet(debug=True)
 
 
 # ❓ FanoutParent
-
-
 class ParentInput(BaseModel):
     n: int = 100
 
@@ -42,13 +40,9 @@ async def spawn(input: ParentInput, ctx: Context) -> dict[str, Any]:
     print(f"results {result}")
 
     return {"results": result}
-
-
 # ‼️
 
 # ❓ FanoutChild
-
-
 @child_wf.task()
 def process(input: ChildInput, ctx: Context) -> dict[str, str]:
     print(f"child process {input.a}")
