+++ sdks/python/examples/simple/worker.py
@@ -1,13 +1,16 @@
 # ❓ Simple
 
-from hatchet_sdk import Context, EmptyModel, Hatchet
+from hatchet_sdk import Context, Hatchet
+from pydantic import BaseModel
 
 hatchet = Hatchet(debug=True)
 
+class SimpleInput(BaseModel):
+    message: str
 
 @hatchet.task(name="SimpleWorkflow")
-def step1(input: EmptyModel, ctx: Context) -> None:
-    print("executed step1")
+def step1(input: SimpleInput, ctx: Context) -> None:
+    print("executed step1: ", input.message)
 
 
 def main() -> None:
